Continued rearranging
