+++ packages/math-expressions-rs/src/calculus/mod.rs
@@ -1,0 +1,6 @@
+//! Calculus: symbolic [`differentiation`](diff) and [`integration`](integrate).
+//!
+//! Future calculus features (limits, series) belong here as sibling modules.
+
+pub mod diff;
+pub mod integrate;
