+++ packages/math-expressions-rs/src/exact/mod.rs
@@ -1,0 +1,27 @@
+//! Exact-constant evaluation and the certified zero-equivalence service.
+//!
+//! Two public entry points:
+//!
+//! * [`Exact`] / [`exact_eval`] — a rigorous evaluator for real constants over
+//!   the field ℚ adjoined with surds (√ of nonnegative rationals), π and e as
+//!   transcendental generators, and the trig/exp/log special values that land
+//!   in that field. It only ever returns a value it can *prove* correct;
+//!   anything outside the tower yields `None`. The ring itself lives in
+//!   [`value`], the evaluator in [`eval`].
+//! * [`is_zero`] — `is_zero(e, a) -> Tri` (`Some(true)` = certified zero,
+//!   `Some(false)` = certified nonzero, `None` = undecided). Soundness is the
+//!   invariant: it never answers `Some(_)` unless the answer is certain. The
+//!   service lives in [`zero`]; the single-`RootOf` decider in [`algebraic`].
+
+mod algebraic;
+mod eval;
+mod value;
+mod zero;
+
+pub use eval::exact_eval;
+pub use value::Exact;
+pub use zero::is_zero;
+
+// Crate-internal entry points (used qualified as `crate::exact::…`).
+pub(crate) use eval::trig_special_value;
+pub(crate) use zero::certified_zero;
