+++ packages/math-expressions-rs/src/ops/mod.rs
@@ -1,0 +1,31 @@
+//! Expression utilities: small, self-contained ports of the corresponding
+//! `me.*` methods, grouped by concern.
+//!
+//! - [`numbers`] — numeric folding and display rounding
+//! - [`query`] — inspection (functions / operators / variables) and component access
+//! - [`transforms`] — structural rewrites (substitute, subscripts, intervals, …)
+//! - [`units`] — unit annotation stripping / adding
+//! - [`analytic`] — the `isAnalytic` predicate
+//! - [`evaluate`] — numeric evaluation at bindings
+//! - [`pm`] — the plus-minus (`±`) operator helpers
+
+mod analytic;
+mod evaluate;
+mod numbers;
+pub mod pm;
+mod query;
+mod transforms;
+mod units;
+
+pub use analytic::{is_analytic, AnalyticOpts};
+pub use evaluate::{evaluate, evaluate_to_constant};
+pub use numbers::{
+    constants_to_floats, evaluate_numbers, reduce_rational, round_numbers_to_decimals,
+    round_numbers_to_precision, round_numbers_to_precision_plus_decimals, set_small_zero,
+};
+pub use query::{functions, get_component, operators, substitute_component, variables};
+pub use transforms::{
+    altvectors_to_vectors, normalize_function_names, strings_to_subscripts, subscripts_to_strings,
+    substitute, to_intervals, tuples_to_vectors,
+};
+pub use units::{add_unit, remove_scaling_units, remove_units};
