+++ packages/math-expressions-rs/src/polynomials/factor.rs
