+++ packages/math-expressions-rs/src/polynomials/rootof.rs
