+++ packages/math-expressions-rs/src/polynomials/univariate.rs
