+++ packages/math-expressions-rs/tests/numeric.rs
@@ -106,44 +106,3 @@ fn check_residual(a: &[f64], n: usize, p: &EigenPair) {
     let mag: f64 = p.vector.iter().map(|v| v.norm_sqr()).sum::<f64>().sqrt();
     assert_close(mag, 1.0, 1e-9);
 }
-
-#[test]
-fn match_template_default_mode() {
-    use math_expressions::js_match::match_template;
-    use serde_json::json;
-
-    // ["+", ["*", 2, "x"], 3] against ["+", ["*", "a", "x"], "b"]:
-    // wildcards a, x, b (all pattern variables).
-    let tree = json!(["+", ["*", 2, "x"], 3]);
-    let pat = json!(["+", ["*", "a", "y"], "b"]);
-    let m = match_template(&tree, &pat).unwrap();
-    assert_eq!(m.get("a").unwrap(), &json!(2));
-    assert_eq!(m.get("y").unwrap(), &json!("x"));
-    assert_eq!(m.get("b").unwrap(), &json!(3));
-
-    // Grouping: last wildcard absorbs the rest of an associative operator.
-    let tree = json!(["+", 1, 2, 3]);
-    let m = match_template(&tree, &json!(["+", "u", "v"])).unwrap();
-    assert_eq!(m.get("u").unwrap(), &json!(1));
-    assert_eq!(m.get("v").unwrap(), &json!(["+", 2, 3]));
-
-    // Repeated wildcard must bind equal subtrees.
-    assert!(match_template(&json!(["+", "x", "x"]), &json!(["+", "u", "u"])).is_some());
-    assert!(match_template(&json!(["+", "x", "y"]), &json!(["+", "u", "u"])).is_none());
-
-    // Unary minus of product matches a * pattern.
-    let tree = json!(["-", ["*", "x", "y"]]);
-    let m = match_template(&tree, &json!(["*", "a", "b"])).unwrap();
-    assert_eq!(m.get("a").unwrap(), &json!(["-", "x"]));
-    assert_eq!(m.get("b").unwrap(), &json!("y"));
-
-    // Operators must match exactly; no match across operators.
-    assert!(match_template(&json!(["*", 1, 2]), &json!(["+", "u", "v"])).is_none());
-    // Exact variable-free match -> empty bindings.
-    assert_eq!(
-        match_template(&json!(["+", 1, 2]), &json!(["+", 1, 2]))
-            .unwrap()
-            .len(),
-        0
-    );
-}
